requirements.txt
python-dotenv
domeneshop
schedule
requests
